get algos working lol
