+++ back/src/VitaWave.Data/DataFacilitator.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Concurrent;
+﻿using Serilog;
+using System.Collections.Concurrent;
+using System.Numerics;
 using System.Reflection.Metadata.Ecma335;
 using VitaWave.Common;
 
@@ -7,14 +9,9 @@ namespace VitaWave.Data
     public class DataFacilitator
     {
         public readonly Dictionary<string, ConcurrentQueue<EventPacket>> _instances = new();
-        public readonly DataProcessor _dataProcessor;
         const int MAX_EVENT_WINDOW = 100; // store the last 100 module events for alg
         private bool dataSave = true;
-
-        public DataFacilitator(DataProcessor dataProcessor)
-        {
-            _dataProcessor = dataProcessor;
-        }
+        public event EventHandler<ResultEvent>? EventRaise;
 
         public void Add(EventPacket packet)
         {
@@ -38,7 +35,7 @@ namespace VitaWave.Data
                             SaveDataHelper.Save(events);
                         }
 
-                        _dataProcessor.NewData(events); // copy to a list instead of pass by ref
+                        OnNewData(events);
                     }
                 }
                 else
@@ -51,6 +48,41 @@ namespace VitaWave.Data
             catch (Exception ex)
             { }
         }
+
+
+        private Dictionary<int, Person> trackedPersons = new(); // key = TID
+        public void OnNewData(List<Common.EventPacket> frames)
+        {
+            foreach (var frame in frames)
+            {
+                double deltaTime = frame.TimeSinceLastMs / 1000.0; // convert to seconds
+
+                if (frame.Targets == null || frame.Targets.Count == 0)
+                    continue;
+
+                foreach (var target in frame.Targets)
+                {
+                    if (!trackedPersons.ContainsKey((int)target.TID))
+                        trackedPersons[(int)target.TID] = new Person();
+
+                    var person = trackedPersons[(int)target.TID];
+                    Vector3 radarPos = new((float)target.X, (float)target.Y, (float)target.Z);
+                    person.UpdatePosition(radarPos, deltaTime);
+                    var posture = person.ClassifyPosture();
+                    posture.TID = (int)target.TID;
+                    Notify(posture);
+                }
+            }
+        }
+
+        public void Notify(ResultEvent e)
+        {
+            Log.Information("Event: " + e.Event + "   " + "TID: " + e.TID);
+
+            //if (EventRaise?.GetInvocationList() != null)
+            //    EventRaise.Invoke(this, e);
+        }
+
 
         public void Clear(string key)
         {
